# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ArabVoice AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Voice {
  id: string;
  name: string;
  language: string;
  gender: string;
  locale: string;
  /** @nullable */
  preview?: string | null;
}

export interface EdgeTtsInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  text: string;
  voiceId: string;
  speed?: number;
  pitch?: number;
  /** @nullable */
  userId?: string | null;
}

export interface XttsInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  text: string;
  speakerAudioUrl: string;
  language?: string;
  /** @nullable */
  userId?: string | null;
}

export interface GenerationResult {
  audioUrl: string;
  engine: string;
  /** Which engine actually produced the audio: railway, hf-fallback, or cached */
  engineUsed?: string;
  /** Whether this result was served from cache (no credits charged) */
  fromCache?: boolean;
  /** @nullable */
  duration: number | null;
  text: string;
  /** @nullable */
  historyId?: string | null;
}

export interface HistoryItem {
  id: string;
  text: string;
  audioUrl: string;
  engine: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  voiceId?: string | null;
  createdAt: string;
}

export interface HistoryPage {
  items: HistoryItem[];
  total: number;
  page: number;
  limit: number;
}

export interface HistoryInput {
  text: string;
  audioUrl: string;
  engine: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  voiceId?: string | null;
  /** @nullable */
  userId?: string | null;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  plan: string;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  avatarUrl?: string;
}

export interface CreditInfo {
  plan: string;
  creditsUsed: number;
  creditsLimit: number;
  /** @nullable */
  resetsAt: string | null;
  canUseXtts?: boolean;
  canDownloadWav?: boolean;
}

export interface UserStats {
  totalGenerations: number;
  todayGenerations: number;
  totalDuration: number;
  /** @nullable */
  favoriteVoice: string | null;
  /** @nullable */
  lastGeneratedAt?: string | null;
}

export interface AdminUser {
  id: string;
  email: string;
  /** @nullable */
  name?: string | null;
  plan: string;
  totalGenerations: number;
  creditsUsed?: number;
  createdAt: string;
}

export interface AdminUsersPage {
  items: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export interface SubscriptionUpdate {
  plan: string;
  /** @nullable */
  creditsLimit?: number | null;
}

export interface AdminStats {
  totalUsers: number;
  totalGenerations: number;
  todayGenerations: number;
  activeSubscriptions: number;
  freeUsers?: number;
  revenueThisMonth?: number;
}

export interface AdminGenerationsPage {
  items: HistoryItem[];
  total: number;
  page: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListHistoryParams = {
page?: number;
limit?: number;
};

export type ListAdminUsersParams = {
page?: number;
limit?: number;
};

export type ListAdminGenerationsParams = {
page?: number;
};