# lib/api-zod/src/generated/types/edgeTtsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ArabVoice AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EdgeTtsInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  text: string;
  voiceId: string;
  speed?: number;
  pitch?: number;
  /** @nullable */
  userId?: string | null;
}