# lib/api-zod/src/generated/types/generationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ArabVoice AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationResult {
  audioUrl: string;
  engine: string;
  /** Which engine actually produced the audio: railway, hf-fallback, or cached */
  engineUsed?: string;
  /** Whether this result was served from cache (no credits charged) */
  fromCache?: boolean;
  /** @nullable */
  duration: number | null;
  text: string;
  /** @nullable */
  historyId?: string | null;
}