# lib/api-zod/src/generated/types/subscriptionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ArabVoice AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubscriptionUpdate {
  plan: string;
  /** @nullable */
  creditsLimit?: number | null;
}