# lib/api-zod/src/generated/types/xttsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ArabVoice AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface XttsInput {
  /**
     * @minLength 1
     * @maxLength 5000
     */
  text: string;
  speakerAudioUrl: string;
  language?: string;
  /** @nullable */
  userId?: string | null;
}